# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CINTEXA CMS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  totalPages: number;
  publishedPages: number;
  draftPages: number;
  totalPosts: number;
  publishedPosts: number;
  totalMedia: number;
  totalUsers: number;
  totalSites: number;
  totalForms: number;
  pageViews: number;
  avgLoadTime: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  entityType: string;
  entityTitle: string;
  userName: string;
  action: string;
  createdAt: string;
}

export interface TrafficPoint {
  date: string;
  views: number;
  visitors: number;
}

export interface Site {
  id: number;
  name: string;
  domain: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  favicon?: string | null;
  /** @nullable */
  logo?: string | null;
  status: string;
  /** @nullable */
  primaryColor?: string | null;
  language: string;
  timezone: string;
  createdAt: string;
  updatedAt: string;
}

export interface SiteInput {
  /** @minLength 1 */
  name: string;
  domain: string;
  description?: string;
  favicon?: string;
  logo?: string;
  primaryColor?: string;
  language?: string;
  timezone?: string;
}

export interface SiteUpdate {
  name?: string;
  domain?: string;
  description?: string;
  favicon?: string;
  logo?: string;
  status?: string;
  primaryColor?: string;
  language?: string;
  timezone?: string;
}

export interface Page {
  id: number;
  /** @nullable */
  siteId?: number | null;
  title: string;
  slug: string;
  status: string;
  /** @nullable */
  template?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  metaTitle?: string | null;
  /** @nullable */
  metaDescription?: string | null;
  /** @nullable */
  featuredImage?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface PageInput {
  siteId?: number;
  /** @minLength 1 */
  title: string;
  slug: string;
  template?: string;
  content?: string;
  metaTitle?: string;
  metaDescription?: string;
  featuredImage?: string;
}

export interface PageUpdate {
  title?: string;
  slug?: string;
  status?: string;
  template?: string;
  content?: string;
  metaTitle?: string;
  metaDescription?: string;
  featuredImage?: string;
}

export interface Post {
  id: number;
  title: string;
  slug: string;
  status: string;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  authorId?: number | null;
  /** @nullable */
  authorName?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  featuredImage?: string | null;
  /** @nullable */
  metaTitle?: string | null;
  /** @nullable */
  metaDescription?: string | null;
  /** @nullable */
  readingTime?: number | null;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface PostInput {
  /** @minLength 1 */
  title: string;
  slug: string;
  excerpt?: string;
  content?: string;
  authorId?: number;
  categoryId?: number;
  featuredImage?: string;
  metaTitle?: string;
  metaDescription?: string;
}

export interface PostUpdate {
  title?: string;
  slug?: string;
  status?: string;
  excerpt?: string;
  content?: string;
  authorId?: number;
  categoryId?: number;
  featuredImage?: string;
  metaTitle?: string;
  metaDescription?: string;
}

export interface MediaItem {
  id: number;
  filename: string;
  originalName: string;
  mimeType: string;
  size: number;
  url: string;
  /** @nullable */
  altText?: string | null;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  createdAt: string;
}

export interface MediaInput {
  /** @minLength 1 */
  filename: string;
  originalName: string;
  mimeType: string;
  size: number;
  url: string;
  altText?: string;
  caption?: string;
  width?: number;
  height?: number;
}

export interface MediaUpdate {
  altText?: string;
  caption?: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  status: string;
  createdAt: string;
}

export interface UserInput {
  /** @minLength 1 */
  name: string;
  email: string;
  role: string;
  avatar?: string;
  bio?: string;
}

export interface UserUpdate {
  name?: string;
  email?: string;
  role?: string;
  avatar?: string;
  bio?: string;
  status?: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  postCount: number;
  createdAt: string;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  slug: string;
  description?: string;
}

export interface CategoryUpdate {
  name?: string;
  slug?: string;
  description?: string;
}

export interface Tag {
  id: number;
  name: string;
  slug: string;
  postCount: number;
  createdAt: string;
}

export interface TagInput {
  /** @minLength 1 */
  name: string;
  slug: string;
}

export interface Menu {
  id: number;
  name: string;
  location: string;
  items: string;
  createdAt: string;
}

export interface MenuInput {
  /** @minLength 1 */
  name: string;
  location: string;
  items: string;
}

export interface MenuUpdate {
  name?: string;
  location?: string;
  items?: string;
}

export interface Form {
  id: number;
  name: string;
  fields: string;
  submissionCount: number;
  createdAt: string;
}

export interface FormInput {
  /** @minLength 1 */
  name: string;
  fields: string;
}

export interface FormUpdate {
  name?: string;
  fields?: string;
}

export interface FormSubmission {
  id: number;
  formId: number;
  data: string;
  createdAt: string;
}

export interface SeoSettings {
  id: number;
  siteTitle: string;
  siteDescription: string;
  robots: string;
  /** @nullable */
  googleAnalyticsId?: string | null;
  /** @nullable */
  googleSearchConsoleId?: string | null;
  /** @nullable */
  ogImage?: string | null;
  /** @nullable */
  twitterHandle?: string | null;
  updatedAt: string;
}

export interface SeoSettingsUpdate {
  siteTitle?: string;
  siteDescription?: string;
  robots?: string;
  googleAnalyticsId?: string;
  googleSearchConsoleId?: string;
  ogImage?: string;
  twitterHandle?: string;
}

export interface Redirect {
  id: number;
  from: string;
  to: string;
  type: number;
  createdAt: string;
}

export interface RedirectInput {
  /** @minLength 1 */
  from: string;
  /** @minLength 1 */
  to: string;
  type: number;
}

export interface AiContentRequest {
  /** @minLength 1 */
  prompt: string;
  type: string;
  tone?: string;
  length?: string;
}

export interface AiContentResponse {
  content: string;
  wordCount: number;
}

export interface AiSeoRequest {
  /** @minLength 1 */
  title: string;
  content: string;
}

export interface AiSeoResponse {
  metaTitle: string;
  metaDescription: string;
  keywords: string[];
}

export interface AiTitleRequest {
  /** @minLength 1 */
  topic: string;
  count?: number;
}

export interface AiTitleResponse {
  titles: string[];
}

export type GetPagesParams = {
siteId?: number;
status?: string;
};

export type GetPostsParams = {
status?: string;
categoryId?: number;
};

export type GetMediaParams = {
type?: string;
search?: string;
};