# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CINTEXA CMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pages:
    get:
      operationId: getPages
      tags: [pages]
      summary: List all pages
      parameters:
        - name: siteId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Pages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Page"
    post:
      operationId: createPage
      tags: [pages]
      summary: Create a page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
  /pages/{id}:
    get:
      operationId: getPage
      tags: [pages]
      summary: Get a page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
    patch:
      operationId: updatePage
      tags: [pages]
      summary: Update a page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
    delete:
      operationId: deletePage
      tags: [pages]
      summary: Delete a page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /pages/{id}/publish:
    post:
      operationId: publishPage
      tags: [pages]
      summary: Publish a page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"

  # ─── Posts ───────────────────────────────────────────────────────────────────
...
  /media:
    get:
      operationId: getMedia
      tags: [media]
      summary: List all media
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Media list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaItem"
    post:
      operationId: createMedia
      tags: [media]
      summary: Upload media (metadata)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaItem"
  /media/{id}:
    get:
      operationId: getMediaItem
      tags: [media]
      summary: Get a media item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Media item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaItem"
    patch:
      operationId: updateMedia
      tags: [media]
      summary: Update media metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaItem"
    delete:
      operationId: deleteMedia
      tags: [media]
      summary: Delete a media item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Users ───────────────────────────────────────────────────────────────────
...
  /menus:
    get:
      operationId: getMenus
      tags: [menus]
      summary: List all menus
      responses:
        "200":
          description: Menus list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Menu"
    post:
      operationId: createMenu
      tags: [menus]
      summary: Create a menu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Menu"
  /menus/{id}:
    patch:
      operationId: updateMenu
      tags: [menus]
      summary: Update a menu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Menu"
    delete:
      operationId: deleteMenu
      tags: [menus]
      summary: Delete a menu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Forms ───────────────────────────────────────────────────────────────────
  /forms:
    get:
      operationId: getForms
      tags: [forms]
      summary: List all forms
      responses:
        "200":
          description: Forms list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Form"
    post:
      operationId: createForm
      tags: [forms]
      summary: Create a form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Form"
  /forms/{id}:
    patch:
      operationId: updateForm
      tags: [forms]
      summary: Update a form
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Form"
    delete:
      operationId: deleteForm
      tags: [forms]
      summary: Delete a form
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /forms/{id}/submissions:
    get:
      operationId: getFormSubmissions
      tags: [forms]
      summary: Get form submissions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FormSubmission"

  # ─── SEO ─────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Page:
      type: object
      properties:
        id:
          type: integer
        siteId:
          type: ["integer", "null"]
        title:
          type: string
        slug:
          type: string
        status:
          type: string
        template:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        featuredImage:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, slug, status, createdAt, updatedAt]
    PageInput:
      type: object
      properties:
        siteId:
          type: integer
        title:
          type: string
          minLength: 1
        slug:
          type: string
        template:
          type: string
        content:
          type: string
        metaTitle:
          type: string
        metaDescription:
          type: string
        featuredImage:
          type: string
      required: [title, slug]
    PageUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        status:
          type: string
        template:
          type: string
        content:
          type: string
        metaTitle:
          type: string
        metaDescription:
          type: string
        featuredImage:
          type: string

    # ─── Post ──────────────────────────────────────────────────────────────────
...
    MediaItem:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        url:
          type: string
        altText:
          type: ["string", "null"]
        caption:
          type: ["string", "null"]
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]
        createdAt:
          type: string
      required: [id, filename, originalName, mimeType, size, url, createdAt]
    MediaInput:
      type: object
      properties:
        filename:
          type: string
          minLength: 1
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        url:
          type: string
        altText:
          type: string
        caption:
          type: string
        width:
          type: integer
        height:
          type: integer
      required: [filename, originalName, mimeType, size, url]
    MediaUpdate:
      type: object
      properties:
        altText:
          type: string
        caption:
          type: string

    # ─── User ──────────────────────────────────────────────────────────────────
...
    Menu:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        items:
          type: string
        createdAt:
          type: string
      required: [id, name, location, items, createdAt]
    MenuInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        location:
          type: string
        items:
          type: string
      required: [name, location, items]
    MenuUpdate:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        items:
          type: string

    # ─── Form ──────────────────────────────────────────────────────────────────
    Form:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        fields:
          type: string
        submissionCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, fields, submissionCount, createdAt]
    FormInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        fields:
          type: string
      required: [name, fields]
    FormUpdate:
      type: object
      properties:
        name:
          type: string
        fields:
          type: string
    FormSubmission:
      type: object
      properties:
        id:
          type: integer
        formId:
          type: integer
        data:
          type: string
        createdAt:
          type: string
      required: [id, formId, data, createdAt]

    # ─── SEO ───────────────────────────────────────────────────────────────────
...