# lib/api-zod/src/generated/types/form.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CINTEXA CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface Form {
  id: number;
  name: string;
  fields: string;
  submissionCount: number;
  createdAt: string;
}